% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTable.R
\name{downloadTable}
\alias{downloadTable}
\alias{downloadTableUI}
\title{Download tables in different formats}
\usage{
downloadTable(input, output, session, rTable, fileName = "tableFile")

downloadTableUI(id)
}
\arguments{
\item{input, output, session}{Standard module parameters}

\item{rTable}{A reactive table (for example a \code{data.frame})}

\item{fileName}{A prefix for the default name in the download handler.}

\item{id}{The module identifier}
}
\value{
Currently, no reactive code is returned due to issues with
\code{downloadTable}. See \href{https://stackoverflow.com/questions/45458348/}{here}
}
\description{
This module uses \code{downloadHandler} to allow downloads of tables in different
formats (\code{xlsx}, \code{rds}, and \code{csv}).
}
\examples{
\dontrun{
shinyApp(
  fluidPage(downloadTableUI("downloadTable")),
  function(input, output, session) {
    callModule(downloadTable, "downloadTable", reactive({mtcars}),
               fileName = "mtcars")
  }
)
}
}
