% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalCode.R
\name{evalCode}
\alias{evalCode}
\title{Evaluate strings with error popups}
\usage{
evalCode(code, env = parent.frame())
}
\arguments{
\item{code}{A \code{character} of length one giving the code to be evaluated.}

\item{env}{The environment where the code should be evaluated. Defaults to
the calling environment.}
}
\value{
The result of the specified expression or \code{NULL} in case of an error
}
\description{
\code{evalCode} tries to evaluate code given as a sting. In case of an error, a
modal pops up showing the resulting error message.
}
\examples{
\dontrun{
shinyApp(
  fluidPage(
    textAreaInput("code", "code", "# intended typo\\nrnrm(100)"),
    actionButton("evaluate", "evaluate"),
    verbatimTextOutput("result")
  ),
  function(input, output, session){
    res <- eventReactive(
      input$evaluate,
      evalCode(input$code)
    )
    output$result <- renderPrint(res())
  }
)
}
}
