% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdownload_popup.R
\name{ggdownload_popup_ui}
\alias{ggdownload_popup_ui}
\alias{ggdownload_popup_server}
\title{Download ggplots with a modal window}
\usage{
ggdownload_popup_ui(id, label = "download...")

ggdownload_popup_server(input, output, session, plotObj)
}
\arguments{
\item{id}{module identifier}

\item{label}{label of the button}

\item{input, output, session}{Standard module parameters}

\item{plotObj}{A reactive ggplot object}
}
\description{
this module invokes a modal window, where parameters for the function
\code{ggsave()} can be specified`
}
\examples{
\dontrun{
shinyApp(
  ui = fluidPage(
    ggdownload_popup_ui("gg_popup")
  ),
  server = function(input, output, session) {
    plot <- reactive({
      library(ggplot2)
      ggplot(mtcars) + aes(wt, mpg) + geom_point()
    })
    callModule(ggdownload_popup_server, "gg_popup", plot)
  }
)
}
}
