% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libData.R
\name{libData}
\alias{libData}
\alias{libDataUI}
\title{Read data from installed packages}
\usage{
libData(input, output, session, assignTo = "dt")

libDataUI(id, selected = "graphSTAT")
}
\arguments{
\item{assignTo}{A character vector of length one. The symbol to assign the data to in the code}
}
\value{
A reactive string representing the import of the dataset. for example
\preformatted{
"data('Arthritis', package = 'vcd')\ndt <- Arthritis"
}
}
\description{
Read data via \code{utils::data} and import it into R. The ui includes a "show documentation" button
to give info about the datasets.
}
\examples{
\dontrun{

shinyApp(
  ui = fluidPage(
    libDataUI("import"),
    verbatimTextOutput("code"),
    dataTableOutput("table")
  ),
  server = function(input, output, session){
    code <- callModule(libData, "import")
    output$code <- renderText(code())
    output$table <- renderDataTable({
      eval(parse(text = code()))
      dt
    })
  }
)

}
}
