% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\alias{readDataUI}
\title{Read data in different formats (xlsx, rds, csv, ...)}
\usage{
readData(input, output, session, path, callback = function() { },
  symbol = NULL)

readDataUI(id)
}
\arguments{
\item{input, output, session}{Standard module parameters.}

\item{path}{A reactive value representing the path of the file to be read.}

\item{callback}{A function that will be called in case of unsupported
formats. Should have no arguments.}

\item{symbol}{Optional: A symbol to use in the import code in place of the
path. This is useful if the path has previously been saved as a variable.}

\item{id}{The module identifier.}
}
\value{
A reactive string representing the code to read the dataset.
}
\description{
Shiny module for reading data into R. currently, the formats \code{.csv}, \code{.sav},
\code{.xlsx}, \code{.rds} and \code{.Rdata} are supported. the following formats will be
mapped \code{.txt -> .csv}, \code{.xls -> .xlsx}.
}
\examples{
\dontrun{

shinyApp(
  fluidPage(
    textInput("path", "choose a path", value = "mtcars.csv"),
    readDataUI("readDataId"),
    codeOutput("code")
  ),
  function(input, output, session){
    code <- callModule(readData, "readDataId", reactive(input$path))
    output$code <- renderCode({code()})
  }
)

}
}
