% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadWidget.R
\name{widgetDownload}
\alias{widgetDownload}
\alias{widgetDownloadUI}
\title{Download widgets as html}
\usage{
widgetDownload(input, output, session, rWidget, filename = "widget.html")

widgetDownloadUI(id, btnText = "Download widget")
}
\arguments{
\item{input, output, session}{Standard shiny module parameters}

\item{rWidget}{A reactive returning a shiny widget}

\item{filename}{Argument to be passed to \link[shiny:downloadHandler]{shiny::downloadHandler}}

\item{id}{The id of the module}

\item{btnText}{The text appearing on the download button}
}
\description{
Shiny module to save shiny widgets as \code{html} files. Internally, the conversion is done with
\link[htmlwidgets:saveWidget]{htmlwidgets::saveWidget}.
}
\examples{
\dontrun{
shinyApp(
  fluidPage(widgetDownloadUI("wd")),
  function(input, output, session) {
    callModule(widgetDownload, "wd", reactive({ DT::datatable(mtcars) }))
  }
)
}
}
