% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggDownload.R
\name{ggDownload}
\alias{ggDownload}
\alias{ggDownloadUI}
\title{Download a ggplot object as an image file.}
\usage{
ggDownload(input, output, session, plotObj, plotObjName = "ggObj")

ggDownloadUI(id, buttonText = "Download image file")
}
\arguments{
\item{input, output, session}{Standard module parameters}

\item{plotObj}{A reactive \code{ggplot} object}

\item{plotObjName}{The name of the object in the resulting code}

\item{id}{The module identifier.}

\item{buttonText}{Text do display on the \code{downloadButton}.}
}
\value{
A reactive string that represents the code for downloading the plot.
}
\description{
Create a ui that contains controls for the height and width of the plot as well as a
\code{downloadButton}. The module uses \link{ggsave} to convert the plot into an image file.
}
\examples{
\dontrun{
library(ggplot2)

gg <- ggplot(mtcars, aes(wt, mpg)) + geom_point()

shinyApp(
  fluidPage(
    ggDownloadUI("ggd"),
    codeOutput("code")
  ),
  function(input, output, session){
    code <- callModule(ggDownload, "ggd", reactive({gg}))
    output$code <- renderCode({
      code()
    })
  }
)
}
}
