% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableView.R
\name{variableView}
\alias{variableView}
\alias{variableViewUI}
\title{SPSS like variable view}
\usage{
variableView(input, output, session, dataset, dataName = "dat")

variableViewUI(id)
}
\arguments{
\item{input, output, session}{Standard module parameters.}

\item{dataset}{A \code{reactive} table (for example a \code{data.frame})}

\item{dataName}{A \code{character} of length one giving the name of the dataset.}

\item{id}{The id of the module.}
}
\value{
A reactive string representing the code to transform the dataset.
}
\description{
This module gives a way to recode dataframes by changing column names and column classes. Also,
datasets can be filtered based on UI inputs.
}
\examples{
\dontrun{
data("tips", package = "reshape2")
## no proper handling for integers yet
shinyApp(
  fluidPage(
    column(6, variableViewUI("vv")),
    column(6, verbatimTextOutput("code"), DTOutput("filtered"))
  ),
  function(input, output, session){
    code <- callModule(variableView, "vv", reactive({tips}))
    output$code <- renderText({ code() })
    output$filtered <- renderDT({
      dat <- tips
      eval(parse(text = code()))
      dat
   })
  },
  options = list(launch.browser = TRUE)
)
}
}
