% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableView.R
\name{variableView}
\alias{variableView}
\alias{variableViewUI}
\alias{selectedVar}
\title{SPSS like variable view}
\usage{
variableView(input, output, session, dataset, dataName = "dat")

variableViewUI(id)

selectedVar(id)
}
\arguments{
\item{input, output, session}{Standard module parameters.}

\item{dataset}{A \code{reactive} table (for example a \code{data.frame})}

\item{dataName}{A \code{character} of length one giving the name of the dataset.}

\item{id}{The id of the module.}
}
\value{
A reactive string representing the code to transform the dataset.
}
\description{
This module gives a way to recode dataframes by changing column names and
column classes. Also, datasets can be filtered based on UI inputs.
}
\details{
\code{selectedVar} is an optional ui element which will show a summary of the
variable selected in
\code{variableViewUI}.
}
\examples{
\dontrun{

library(DT)
data("tips", package = "reshape2")
data("diamonds", package = "ggplot2")

shinyApp(
  fluidPage(
    column(6, selectInput("dataset", "choose dataset",
           choices = c("mtcars", "tips", "diamonds")),
              variableViewUI("vv")),
    column(6, codeOutput("code"), selectedVar("vv"), DTOutput("filtered"))
  ),
  function(input, output, session){
    dataset <- reactive({get(input$dataset)})

    code <- callModule(variableView, "vv", dataset)
    output$code <- renderCode({ code() })
    output$filtered <- renderDT({
      dat <- isolate(dataset())
      eval(parse(text = code()))
      dat
   })
  },
  options = list(launch.browser = TRUE)
)
}
}
